requirements.txt
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
tensorflow==2.15.0
keras==2.15.0
ta-lib==0.4.28
gymnasium==0.29.1
stable-baselines3==2.2.1
ccxt==4.1.9
python-binance==1.0.19
websockets==12.0
fastapi==0.109.0
uvicorn==0.25.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
jupyter==1.0.0
notebook==7.0.6
ipywidgets==8.1.1
tqdm==4.66.1